+++ src/ReadonlyNonEmptyArray.ts
@@ -21,7 +21,7 @@ import { Eq, fromEquals } from './Eq'
 import { Extend1 } from './Extend'
 import { Foldable1 } from './Foldable'
 import { FoldableWithIndex1 } from './FoldableWithIndex'
-import { identity, Lazy, pipe, Predicate, Refinement } from './function'
+import { Endomorphism, identity, Lazy, pipe, Predicate, Refinement } from './function'
 import { bindTo as bindTo_, flap as flap_, Functor1 } from './Functor'
 import { FunctorWithIndex1 } from './FunctorWithIndex'
 import { HKT } from './HKT'
@@ -1117,7 +1117,7 @@ export const max = <A>(O: Ord<A>): ((as: ReadonlyNonEmptyArray<A>) => A) => {
 export const concatAll = <A>(S: Semigroup<A>) => (as: ReadonlyNonEmptyArray<A>): A => as.reduce(S.concat)
 
 /**
- * Break an array into its first element and remaining elements
+ * Break a `ReadonlyArray` into its first element and remaining elements.
  *
  * @category destructors
  * @since 2.11.0
@@ -1126,7 +1126,7 @@ export const matchLeft = <A, B>(f: (head: A, tail: ReadonlyArray<A>) => B) => (a
   f(head(as), tail(as))
 
 /**
- * Break an array into its initial elements and the last element
+ * Break a `ReadonlyArray` into its initial elements and the last element.
  *
  * @category destructors
  * @since 2.11.0
@@ -1135,22 +1135,22 @@ export const matchRight = <A, B>(f: (init: ReadonlyArray<A>, last: A) => B) => (
   f(init(as), last(as))
 
 /**
- * Modifies the first element of the array
+ * Apply a function to the head, creating a new `ReadonlyNonEmptyArray`.
  *
  * @since 2.11.0
  */
-export const modifyHead = <A>(f: (a: A) => A): ((nea: ReadonlyNonEmptyArray<A>) => ReadonlyNonEmptyArray<A>) => {
-  return matchLeft((head, tail) => pipe(tail, prepend(f(head))))
-}
+export const modifyHead = <A>(f: Endomorphism<A>) => (as: ReadonlyNonEmptyArray<A>): ReadonlyNonEmptyArray<A> => [
+  f(head(as)),
+  ...tail(as)
+]
 
 /**
- * Modifies the last element of the array
+ * Apply a function to the last element, creating a new `ReadonlyNonEmptyArray`.
  *
  * @since 2.11.0
  */
-export const modifyLast = <A>(f: (a: A) => A): ((nea: ReadonlyNonEmptyArray<A>) => ReadonlyNonEmptyArray<A>) => {
-  return matchRight((init, last) => pipe(init, append(f(last))))
-}
+export const modifyLast = <A>(f: Endomorphism<A>) => (as: ReadonlyNonEmptyArray<A>): ReadonlyNonEmptyArray<A> =>
+  pipe(init(as), append(f(last(as))))
 
 // -------------------------------------------------------------------------------------
 // deprecated